refactor pre setuptools
