+++ lib/linodefsstat.py
@@ -1,0 +1,19 @@
+try:
+    import _find_fuse_parts
+except ImportError:
+    pass
+import fuse
+
+class LinodeFSStat(fuse.Stat):
+
+    def __init__(self):
+        self.st_mode = 0
+        self.st_ino = 0
+        self.st_dev = 0
+        self.st_nlink = 0
+        self.st_uid = 0
+        self.st_gid = 0
+        self.st_size = 0
+        self.st_atime = 0
+        self.st_mtime = 0
+        self.st_ctime = 0
