+++ linodefs/__init__.py
@@ -1,0 +1,4 @@
+import app main
+
+if __name__ == '__main__':
+    main()
